Refactor allowMemberInRoom controller using RoomMember type classifier functions
* The helper functions were further abstracted and split up, to honor
  the single responsibility principle.

* Helper functions were added for the room member, that helps us
  distinguish whether the member is connected, allowed etc. These were
  used wherever they're needed in the project.
