+++ src/lambdas/roomWsAuthorizer/index.production.ts
@@ -1,6 +1,7 @@
 import { createConnection } from "mongoose";
 import { APIGatewayRequestAuthorizerHandler as Handler } from "aws-lambda";
 import makeRoomModel from "models/Room/index.model";
+import { getIsMemberConnected } from "models/Room/schemas/member/helperFns";
 import ROOM_WS_AUTHORIZER_ENV_VARS from "./env";
 import generateAuthPolicy from "./utils/generateAuthPolicy";
 import parseEventData from "./utils/eventData";
@@ -23,24 +24,33 @@ export const handler: Handler = async ({
     requestContext,
   });
   if (!eventDataParseResult.success) return rejectedAuthPolicy;
-  const { connectionId, roomId, memberId, nonce } = eventDataParseResult.data;
+  const {
+    connectionId,
+    roomId,
+    memberId: requestingMemberId,
+    nonce,
+  } = eventDataParseResult.data;
 
   const Room = makeRoomModel(mongoClient);
   const requestedRoom = await Room.findById(roomId);
   if (!requestedRoom) return rejectedAuthPolicy;
 
   // make sure that the room is ready to accept members
-  const isThisRequestNotFromAdmin = requestedRoom.adminMemberId !== memberId;
-  const isAdminNotConnected = !requestedRoom.getAdminMember().connectionId;
+  const adminMember = requestedRoom.getAdminMember();
+  const isThisRequestNotFromAdmin = adminMember.memberId !== requestingMemberId;
+  const isAdminNotConnected = !getIsMemberConnected(adminMember);
   if (isThisRequestNotFromAdmin && isAdminNotConnected)
     return rejectedAuthPolicy;
 
   // process the sender of this request
-  const isReqVerified = requestedRoom.verifyNonce(memberId, nonce);
+  const isReqVerified = requestedRoom.verifyNonce(requestingMemberId, nonce);
   if (!isReqVerified) return rejectedAuthPolicy;
-  await requestedRoom.confirmConnection(memberId, connectionId);
+  await requestedRoom.confirmConnection(requestingMemberId, connectionId);
   if (isThisRequestNotFromAdmin)
-    await new AdminNotifier(memberId, requestedRoom).notifyAboutNewMember();
+    await new AdminNotifier(
+      requestingMemberId,
+      requestedRoom
+    ).notifyAboutNewMember();
 
   return generateAuthPolicy(true, methodArn);
 };
